social_media_api/requirements.txt
django
djangorestframework
djangorestframework-authtoken
django-filter
Pillow
